+++ eras/conway/test-suite/test/Test/Cardano/Ledger/Conway/Serialisation/CDDL.hs
@@ -12,8 +12,10 @@ import Cardano.Ledger.Alonzo.Tx (Data)
 import Cardano.Ledger.Alonzo.TxWits (Redeemers)
 import Cardano.Ledger.Babbage.TxBody (Datum)
 import Cardano.Ledger.Conway (Conway)
+import Cardano.Ledger.Conway.Delegation.Certificates (ConwayDCert)
 import Cardano.Ledger.Conway.Governance (ProposalProcedure, VotingProcedure)
 import Cardano.Ledger.Core
+import Cardano.Ledger.Crypto (StandardCrypto)
 import qualified Data.ByteString.Lazy as BSL
 import Paths_cardano_ledger_conway_test (getDataFileName)
 import Test.Cardano.Ledger.Shelley.Serialisation.CDDLUtils (
@@ -41,6 +43,7 @@ tests n = withResource combinedCDDL (const (pure ())) $ \cddl ->
     , cddlAnnotatorTest @(Tx Conway) (eraProtVerHigh @Conway) n "transaction"
     , cddlTest @(VotingProcedure Conway) (eraProtVerHigh @Conway) n "voting_procedure"
     , cddlTest @(ProposalProcedure Conway) (eraProtVerHigh @Conway) n "proposal_procedure"
+    , cddlTest @(ConwayDCert StandardCrypto) (eraProtVerHigh @Conway) n "certificate"
     ]
       <*> pure cddl